refactored alarm tools and all modules except ec2-modules to use the new alarm tools. This cleans up substantial repeated code and makes each module much shorter and much more maintainable. Core logic remains the same.
