+++ handlers/src/alarm-tools.mts
@@ -3,6 +3,7 @@ import {
   ComparisonOperator,
   DeleteAlarmsCommand,
   DescribeAlarmsCommand,
+  MetricDataQuery,
   PutAnomalyDetectorCommand,
   PutMetricAlarmCommand,
   Statistic,
@@ -18,7 +19,11 @@ import {
   DescribeWorkspaceCommand,
   DescribeWorkspaceCommandInput,
 } from '@aws-sdk/client-amp';
-import {MissingDataTreatment} from './alarm-config.mjs';
+import {
+  MetricAlarmConfig,
+  MetricAlarmOptions,
+  MissingDataTreatment,
+} from './alarm-config.mjs';
 import {ConfiguredRetryStrategy} from '@smithy/util-retry';
 import * as yaml from 'js-yaml';
 import * as aws4 from 'aws4';
@@ -56,6 +61,458 @@ export async function doesAlarmExist(alarmName: string): Promise<boolean> {
   return (response.MetricAlarms?.length ?? 0) > 0;
 }
 
+export async function deleteExistingAlarms(
+  service: string,
+  identifier: string,
+) {
+  log
+    .info()
+    .str('function', 'deleteExistingAlarms')
+    .str('Service', service)
+    .str('Identifier', identifier)
+    .msg('Fetching and deleting existing alarms');
+  const activeAutoAlarms = await getCWAlarmsForInstance(service, identifier);
+
+  log
+    .info()
+    .str('function', 'deleteExistingAlarms')
+    .obj('AlarmName', activeAutoAlarms)
+    .msg('Deleting alarm');
+  await cloudWatchClient.send(
+    new DeleteAlarmsCommand({
+      AlarmNames: [...activeAutoAlarms],
+    }),
+  );
+}
+
+async function deleteAlarmsForConfig(
+  config: MetricAlarmConfig,
+  serviceIdentifier: string,
+) {
+  for (const classification of Object.values(AlarmClassification)) {
+    for (const alarmVariant of ['static', 'anomaly'] as const) {
+      const alarmName = buildAlarmName(
+        config,
+        serviceIdentifier,
+        classification,
+        alarmVariant,
+      );
+      await deleteAlarm(alarmName);
+    }
+  }
+}
+
+async function deleteAlarm(alarmName: string) {
+  log
+    .info()
+    .str('function', 'deleteAlarm')
+    .str('AlarmName', alarmName)
+    .msg('Attempting to delete alarm');
+  try {
+    await cloudWatchClient.send(
+      new DeleteAlarmsCommand({AlarmNames: [alarmName]}),
+    );
+    log
+      .info()
+      .str('function', 'deleteAlarm')
+      .str('AlarmName', alarmName)
+      .msg('Successfully deleted alarm');
+  } catch (e) {
+    log
+      .error()
+      .str('function', 'deleteAlarm')
+      .str('AlarmName', alarmName)
+      .err(e)
+      .msg('Error deleting alarm');
+  }
+}
+
+export function buildAlarmName(
+  config: MetricAlarmConfig,
+  serviceIdentifier: string,
+  classification: AlarmClassification,
+  alarmVarient: 'anomaly' | 'static',
+) {
+  const alarmName =
+    alarmVarient === 'anomaly'
+      ? `AutoAlarm-ALB-${serviceIdentifier}-${config.metricName}-anomaly-${classification}`
+      : `AutoAlarm-ALB-${serviceIdentifier}-${config.metricName}-${classification}`;
+  log
+    .info()
+    .str('function', 'buildAlarmName')
+    .str('AlarmName', alarmName)
+    .msg('Built alarm name name');
+  return alarmName;
+}
+
+async function handleAnomalyDetectionWorkflow(
+  alarmName: string,
+  updatedDefaults: MetricAlarmOptions,
+  config: MetricAlarmConfig,
+  dimensions: {Name: string; Value: string}[],
+  classification: AlarmClassification,
+  threshold: number,
+) {
+  log
+    .info()
+    .str('function', 'handleAnomalyDetectionWorkflow')
+    .str('AlarmName', alarmName)
+    .msg('Handling anomaly detection alarm workflow');
+
+  const anomalyDetectorInput = {
+    Namespace: config.metricNamespace,
+    MetricName: config.metricName,
+    Dimensions: [...dimensions],
+    Stat: updatedDefaults.statistic,
+    Configuration: {MetricTimezone: 'UTC'},
+  };
+
+  log
+    .debug()
+    .str('function', 'handleAnomalyDetectionWorkflow')
+    .obj('AnomalyDetectorInput', anomalyDetectorInput)
+    .msg('Sending PutAnomalyDetectorCommand');
+  const response = await cloudWatchClient.send(
+    new PutAnomalyDetectorCommand(anomalyDetectorInput),
+  );
+  log
+    .info()
+    .str('function', 'handleAnomalyDetectionWorkflow')
+    .str('AlarmName', alarmName)
+    .obj('response', response)
+    .msg('Successfully created or updated anomaly detector');
+
+  const metrics: MetricDataQuery[] = [
+    {
+      Id: 'primaryMetric',
+      MetricStat: {
+        Metric: {
+          Namespace: config.metricNamespace,
+          MetricName: config.metricName,
+          Dimensions: [...dimensions],
+        },
+        Period: updatedDefaults.period,
+        Stat: updatedDefaults.statistic,
+      },
+    },
+    {
+      Id: 'anomalyDetectionBand',
+      Expression: `ANOMALY_DETECTION_BAND(primaryMetric, ${threshold})`,
+    },
+  ];
+
+  try {
+    const alarmInput = {
+      AlarmName: alarmName,
+      ComparisonOperator: updatedDefaults.comparisonOperator,
+      EvaluationPeriods: updatedDefaults.evaluationPeriods,
+      Metrics: metrics,
+      ThresholdMetricId: 'anomalyDetectionBand',
+      ActionsEnabled: false,
+      Tags: [{Key: 'severity', Value: classification}],
+      TreatMissingData: updatedDefaults.missingDataTreatment,
+    };
+
+    log
+      .info()
+      .str('function', 'handleAnomalyDetectionWorkflow')
+      .obj('AlarmInput', alarmInput)
+      .msg('Sending PutMetricAlarmCommand');
+
+    const response = await cloudWatchClient.send(
+      new PutMetricAlarmCommand(alarmInput),
+    );
+    log
+      .info()
+      .str('function', 'handleAnomalyDetectionWorkflow')
+      .str('AlarmName', alarmName)
+      .obj('response', response)
+      .msg('Successfully created or updated anomaly detection alarm');
+  } catch (e) {
+    log
+      .error()
+      .str('function', 'handleAnomalyDetectionWorkflow')
+      .str('AlarmName', alarmName)
+      .err(e)
+      .msg('Error creating or updating anomaly detection alarm');
+  }
+}
+
+export async function handleAnomalyAlarms(
+  config: MetricAlarmConfig,
+  serviceIdentifier: string,
+  dimensions: {Name: string; Value: string}[],
+  updatedDefaults: MetricAlarmOptions,
+): Promise<string[]> {
+  const createdAlarms: string[] = [];
+
+  // Validate if thresholds are set correctly
+  const warningThresholdSet =
+    updatedDefaults.warningThreshold !== undefined &&
+    updatedDefaults.warningThreshold !== null;
+  const criticalThresholdSet =
+    updatedDefaults.criticalThreshold !== undefined &&
+    updatedDefaults.criticalThreshold !== null;
+
+  // If no thresholds are set, log and exit early
+  if (!warningThresholdSet && !criticalThresholdSet && !config.defaultCreate) {
+    const alarmPrefix = `AutoAlarm-ALB-${serviceIdentifier}-${config.metricName}-anomaly-`;
+    log
+      .info()
+      .str('function', 'handleAnomalyAlarms')
+      .str('Service Identifier', serviceIdentifier)
+      .str('alarm prefix: ', alarmPrefix)
+      .msg(
+        'No thresholds defined, skipping alarm creation and deleting alarms for config if they exist.',
+      );
+    await deleteAlarmsForConfig(config, serviceIdentifier);
+    return createdAlarms;
+  }
+
+  // Handle warning anomaly alarm
+  if (warningThresholdSet) {
+    const warningAlarmName = buildAlarmName(
+      config,
+      serviceIdentifier,
+      AlarmClassification.Warning,
+      'anomaly',
+    );
+    log
+      .info()
+      .str('function', 'handleAnomalyAlarms')
+      .str('AlarmName', warningAlarmName)
+      .msg('Creating or updating warning anomaly alarm');
+    await handleAnomalyDetectionWorkflow(
+      warningAlarmName,
+      updatedDefaults,
+      config,
+      dimensions,
+      AlarmClassification.Warning,
+      updatedDefaults.warningThreshold as number,
+    );
+    createdAlarms.push(warningAlarmName);
+  } else {
+    const warningAlarmName = buildAlarmName(
+      config,
+      serviceIdentifier,
+      AlarmClassification.Warning,
+      'anomaly',
+    );
+    log
+      .info()
+      .str('function', 'handleAnomalyAlarms')
+      .str('AlarmName', warningAlarmName)
+      .msg('Deleting existing warning anomaly alarm due to no threshold.');
+    await deleteAlarm(warningAlarmName);
+  }
+
+  // Handle critical anomaly alarm
+  if (criticalThresholdSet) {
+    const criticalAlarmName = buildAlarmName(
+      config,
+      serviceIdentifier,
+      AlarmClassification.Critical,
+      'anomaly',
+    );
+    log
+      .info()
+      .str('function', 'handleAnomalyAlarms')
+      .str('AlarmName', criticalAlarmName)
+      .msg('Creating or updating critical anomaly alarm');
+    await handleAnomalyDetectionWorkflow(
+      criticalAlarmName,
+      updatedDefaults,
+      config,
+      dimensions,
+      AlarmClassification.Critical,
+      updatedDefaults.criticalThreshold as number,
+    );
+    createdAlarms.push(criticalAlarmName);
+  } else {
+    const criticalAlarmName = buildAlarmName(
+      config,
+      serviceIdentifier,
+      AlarmClassification.Critical,
+      'anomaly',
+    );
+    log
+      .info()
+      .str('function', 'handleAnomalyAlarms')
+      .str('AlarmName', criticalAlarmName)
+      .msg('Deleting existing critical anomaly alarm due to no threshold.');
+    await deleteAlarm(criticalAlarmName);
+  }
+
+  return createdAlarms;
+}
+
+async function handleStaticThresholdWorkflow(
+  alarmName: string,
+  updatedDefaults: MetricAlarmOptions,
+  config: MetricAlarmConfig,
+  dimensions: {Name: string; Value: string}[],
+  classification: AlarmClassification,
+  threshold: number,
+) {
+  log
+    .info()
+    .str('function', 'handleStaticThresholdWorkflow')
+    .str('AlarmName', alarmName)
+    .msg('Handling static threshold alarm workflow');
+
+  try {
+    const alarmInput = {
+      AlarmName: alarmName,
+      ComparisonOperator: updatedDefaults.comparisonOperator,
+      EvaluationPeriods: updatedDefaults.evaluationPeriods,
+      MetricName: config.metricName,
+      Namespace: config.metricNamespace,
+      Period: updatedDefaults.period,
+      ...(['p', 'tm', 'tc', 'ts', 'wm', 'iqm'].some((prefix) =>
+        updatedDefaults.statistic.startsWith(prefix),
+      )
+        ? {ExtendedStatistic: updatedDefaults.statistic}
+        : {Statistic: updatedDefaults.statistic as Statistic}),
+      Threshold: threshold,
+      ActionsEnabled: false,
+      Dimensions: [...dimensions],
+      Tags: [{Key: 'severity', Value: classification}],
+      TreatMissingData: updatedDefaults.missingDataTreatment,
+    };
+
+    log
+      .debug()
+      .str('function', 'handleStaticThresholdWorkflow')
+      .obj('AlarmInput', alarmInput)
+      .msg('Sending PutMetricAlarmCommand');
+    const response = await cloudWatchClient.send(
+      new PutMetricAlarmCommand(alarmInput),
+    );
+    log
+      .info()
+      .str('function', 'handleStaticThresholdWorkflow')
+      .str('AlarmName', alarmName)
+      .obj('response', response)
+      .msg('Successfully created or updated static threshold alarm');
+  } catch (e) {
+    log
+      .error()
+      .str('function', 'handleStaticThresholdWorkflow')
+      .str('AlarmName', alarmName)
+      .err(e)
+      .msg('Error creating or updating static threshold alarm');
+  }
+}
+
+export async function handleStaticAlarms(
+  config: MetricAlarmConfig,
+  serviceIdentifier: string,
+  dimensions: {Name: string; Value: string}[],
+  updatedDefaults: MetricAlarmOptions,
+): Promise<string[]> {
+  const createdAlarms: string[] = [];
+
+  // Validate if thresholds are set correctly
+  const warningThresholdSet =
+    updatedDefaults.warningThreshold !== undefined &&
+    updatedDefaults.warningThreshold !== null;
+  const criticalThresholdSet =
+    updatedDefaults.criticalThreshold !== undefined &&
+    updatedDefaults.criticalThreshold !== null;
+
+  // If no thresholds are set, log and exit early
+  if (!warningThresholdSet && !criticalThresholdSet && !config.defaultCreate) {
+    const alarmPrefix = `AutoAlarm-ALB-${serviceIdentifier}-${config.metricName}`;
+    log
+      .info()
+      .str('function', 'handleStaticAlarms')
+      .str('serviceIdentifier', serviceIdentifier)
+      .str('alarm prefix: ', `${alarmPrefix}`)
+      .msg(
+        'No thresholds defined, skipping alarm creation and deleting alarms for config if they exist.',
+      );
+    await deleteAlarmsForConfig(config, serviceIdentifier);
+    return createdAlarms;
+  }
+
+  // Handle warning static alarm
+  if (warningThresholdSet) {
+    const warningAlarmName = buildAlarmName(
+      config,
+      serviceIdentifier,
+      AlarmClassification.Warning,
+      'static',
+    );
+    log
+      .info()
+      .str('function', 'handleStaticAlarms')
+      .str('AlarmName', warningAlarmName)
+      .msg('Creating or updating warning static alarms');
+    await handleStaticThresholdWorkflow(
+      warningAlarmName,
+      updatedDefaults,
+      config,
+      dimensions,
+      AlarmClassification.Warning,
+      updatedDefaults.warningThreshold as number,
+    );
+    createdAlarms.push(warningAlarmName);
+  } else {
+    const warningAlarmName = buildAlarmName(
+      config,
+      serviceIdentifier,
+      AlarmClassification.Warning,
+      'static',
+    );
+    log
+      .info()
+      .str('function', 'handleStaticAlarms')
+      .str('AlarmName', warningAlarmName)
+      .msg('Deleting existing warning static alarm due to no threshold.');
+    await deleteAlarm(warningAlarmName);
+  }
+
+  // Handle critical static alarm
+  if (criticalThresholdSet) {
+    const criticalAlarmName = buildAlarmName(
+      config,
+      serviceIdentifier,
+      AlarmClassification.Critical,
+      'static',
+    );
+    log
+      .info()
+      .str('function', 'handleStaticAlarms')
+      .str('AlarmName', criticalAlarmName)
+      .msg('Creating or updating critical static alarms');
+    await handleStaticThresholdWorkflow(
+      criticalAlarmName,
+      updatedDefaults,
+      config,
+      dimensions,
+      AlarmClassification.Critical,
+      updatedDefaults.criticalThreshold as number,
+    );
+    createdAlarms.push(criticalAlarmName);
+  } else {
+    const criticalAlarmName = buildAlarmName(
+      config,
+      serviceIdentifier,
+      AlarmClassification.Critical,
+      'static',
+    );
+    log
+      .info()
+      .str('function', 'handleStaticAlarms')
+      .str('AlarmName', criticalAlarmName)
+      .msg('Deleting existing critical static alarm due to no threshold.');
+    await deleteAlarm(criticalAlarmName);
+  }
+
+  return createdAlarms;
+}
+
+// the following alarm udpate function will be later deleted and deprecated in favor of the functions above.
 // returns true if the alarm needs to be updated whether it exists or does not. For Anomaly Detection CW alarms
 export async function anomalyCWAlarmNeedsUpdate(
   alarmName: string,
